added nav bar. changed file structure to /src
