+++ src/styles/Navigation.module.scss
@@ -1,0 +1,11 @@
+.MainNav {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: #420543;
+  color: white;
+}
+
+.NavItem {
+  color: white;
+}
